requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
atlastk==0.13.4
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
idna==3.10
iso8601==2.1.0
magic-filter==1.0.12
multidict==6.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pypika-tortoise==0.5.0
python-dotenv==1.0.1
pytz==2025.1
sniffio==1.3.1
starlette==0.46.1
tortoise==0.1.1
tortoise-orm==0.24.2
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3

redis~=5.2.1
cryptography~=44.0.2